// AngularMaterialApp/src/app/model/model.component.ts
export class ModelComponent {

  name?: string = '';
  email?: string = '';
  password?: string = 'null';
  gender?: string = 'Male';
  date?: string = '';
  mobile?: string = '';
  skill?: string = 'null';
  fname?: string = '';
  name1?: string = 'default';
  mname?: string = '';
  bankname?: string = '';
  bankaccountno?: string = 'default';
  ifsccode?: string = 'default';
  mobilebanking?: string = 'default';
  internetbanking?: string = 'default';
  presentaddress?: string;
  permanentaddress?: string;

}